+++ spring-api-services/src/main/java/io/riggo/web/LoadStopController.java
@@ -64,7 +64,7 @@ public class LoadStopController {
                     BaseAPIResponse<LoadStop> loadStopBaseAPIResponse = new BaseAPIResponse<>();
                     if(StringUtils.isNotBlank(loadStop.getLoadExtSysId())) {
                         Optional<Load> optionalLoad = loadService.findByExtSysId(loadStop.getLoadExtSysId(), authenticationFacade.getSiteId());
-                        if (optionalLoad.isPresent() && (loadId == null || optionalLoad.get().getId() == loadId)) {
+                        if (optionalLoad.isPresent() && (loadId == null || optionalLoad.get().getId().intValue() == loadId.intValue())) {
                             if (StringUtils.isNotBlank(loadStop.getExtSysId())) {
                                 Optional<LoadStop> loadStopFromDb = loadStopService.findByExtSysId(loadStop.getExtSysId(), authenticationFacade.getSiteId());
                                 if (loadStopFromDb.isPresent()) {
